uniform use as_dict
