+++ camel/configs/groq_config.py
@@ -13,7 +13,7 @@
 # =========== Copyright 2023 @ CAMEL-AI.org. All Rights Reserved. ===========
 from __future__ import annotations
 
-from typing import Any, Optional, Sequence, Union
+from typing import Optional, Sequence, Union
 
 from openai._types import NOT_GIVEN, NotGiven
 
@@ -102,14 +102,6 @@ class GroqConfig(BaseConfig):
     user: str = ""
     tools: Optional[list[OpenAIFunction]] = None
     tool_choice: Optional[Union[dict[str, str], str]] = "none"
-
-    def as_dict(self) -> dict[str, Any]:
-        config_dict = self.model_dump()
-        if self.tools:
-            config_dict["tools"] = [
-                tool.get_openai_tool_schema() for tool in self.tools
-            ]
-        return config_dict
 
 
 GROQ_API_PARAMS = {param for param in GroqConfig.model_fields.keys()}
