+++ examples/function_call/code_execution.py
@@ -35,7 +35,7 @@ assistant_model_config = ChatGPTConfig(
 model = ModelFactory.create(
     model_platform=ModelPlatformType.OPENAI,
     model_type=ModelType.GPT_3_5_TURBO,
-    model_config_dict=assistant_model_config.model_dump(),
+    model_config_dict=assistant_model_config.as_dict(),
 )
