+++ examples/generate_text_embedding_data/task_generation.py
@@ -38,7 +38,7 @@ def main() -> None:
     model = ModelFactory.create(
         model_platform=ModelPlatformType.OPENAI,
         model_type=ModelType.GPT_3_5_TURBO,
-        model_config_dict=ChatGPTConfig(temperature=0.0).model_dump(),
+        model_config_dict=ChatGPTConfig(temperature=0.0).as_dict(),
     )
     agent = ChatAgent(
         assistant_sys_msg,
