+++ examples/misalignment/role_playing_multiprocess.py
@@ -48,7 +48,7 @@ def generate_data(
             model=ModelFactory.create(
                 model_platform=ModelPlatformType.OPENAI,
                 model_type=ModelType.GPT_3_5_TURBO,
-                model_config_dict=ChatGPTConfig(temperature=1.4).model_dump(),
+                model_config_dict=ChatGPTConfig(temperature=1.4).as_dict(),
             )
         ),
     )
