+++ examples/models/azure_openai_model_example.py
@@ -28,7 +28,7 @@ export AZURE_DEPLOYMENT_NAME=""
 model = ModelFactory.create(
     model_platform=ModelPlatformType.AZURE,
     model_type=ModelType.GPT_3_5_TURBO,
-    model_config_dict=ChatGPTConfig(temperature=0.2).model_dump(),
+    model_config_dict=ChatGPTConfig(temperature=0.2).as_dict(),
 )
 
 # Define system message
