+++ examples/models/role_playing_with_claude.py
@@ -26,7 +26,7 @@ def main(model_type=None) -> None:
     model = ModelFactory.create(
         model_platform=ModelPlatformType.ANTHROPIC,
         model_type=model_type,
-        model_config_dict=AnthropicConfig().model_dump(),
+        model_config_dict=AnthropicConfig().as_dict(),
     )
 
     # Update agent_kwargs to use the created models
