+++ examples/models/role_playing_with_gemini.py
@@ -27,7 +27,7 @@ def main(model_type=None) -> None:
     model = ModelFactory.create(
         model_platform=ModelPlatformType.GEMINI,
         model_type=model_type,
-        model_config_dict=GeminiConfig().model_dump(),
+        model_config_dict=GeminiConfig().as_dict(),
     )
 
     # Update agent_kwargs to use the created models
