+++ examples/models/zhipuai_model_example.py
@@ -21,7 +21,7 @@ from camel.types import ModelPlatformType, ModelType
 model = ModelFactory.create(
     model_platform=ModelPlatformType.ZHIPU,
     model_type=ModelType.GLM_4,
-    model_config_dict=ZhipuAIConfig(temperature=0.2).model_dump(),
+    model_config_dict=ZhipuAIConfig(temperature=0.2).as_dict(),
 )
 
 # Define system message
